cricket_draft_bot/requirements.txt
python-telegram-bot>=20.0
requests
tzlocal<5.0
google-generativeai
beautifulsoup4
python-dotenv
pymongo
dnspython
